rename: 홈 컴포넌트 디렉터리 이동
